APP-2031 - Thread leak fix
Integration Bridge was leaking threads and making the application itself
unresponsible, preventing it to answer any integration requests.
This changeset introduces threading fix in two fronts:

1) Core - There were 10 scheduled threads trying to verify apps
bootstrap. This was changed to have one scheduler with one thread to
handle the apss bootstrap. The schedures, before, were sending the
bootstraps to other thread pool. This model was changed to not have the
former thread pool. All the bootstrap work is now performed in the
scheduler thread.

2) HealthCheck - The health checks were using the async http client of
Jersey implementation, but the async threads created were just blocked
in the next line. The http client call was reverted to a blocking call
and explicitly closed after its execution.